+++ Installomator app script.sh
@@ -17,7 +17,7 @@ if [ ! -e "${destFile}" ]; then
     exit 99
 fi
 
-${destFile} ${what} #BLOCKING_PROCESS_ACTION=tell_user_then_kill
+${destFile} ${what} #BLOCKING_PROCESS_ACTION=tell_user_then_kill #NOTIFY=all
 if [ $? != 0 ]; then
 # This is currently not working in Mosyle, that will ignore script errors. Please request support for this from Mosyle!
     echo "Error installing ${what}. Exit code $?"
@@ -27,6 +27,14 @@ fi
 echo "[$(DATE)][LOG-END]"
 
 exit 0
+
+# notify behavior
+# Default is
+# NOTIFY=success
+# options:
+#   - success      notify the user on success
+#   - silent       no notifications
+#   - all          all notifications (great for Self Service installation)
 
 # behavior when blocking processes are found
 # Default is